backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deprecation==2.1.0
fastapi==0.115.1
frozenlist==1.4.1
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
gotrue==2.9.2
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
packaging==24.1
postgrest==0.17.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
realtime==2.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==13.1
yarl==1.15.0
